Merge [Student] into [develop]

+++ Uspihh.Models/EntityModels/SubjectEntity.cs
@@ -4,7 +4,7 @@ namespace Uspihh.Models.EntityModels
 {
     public class SubjectEntity
     {
-        public long SubjectId { get; set; }
+        public short SubjectId { get; set; }
         public string SubjectName { get; set; }
         public IList<GroupEntity> Groups { get; set; }
     }
